+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Change history for ui-finc-select
 
-## 6.2.0 (IN PROGRESS)
+## 7.0.0 (IN PROGRESS)
 * Use functional components instead of class components ([UIFC-360](https://folio-org.atlassian.net/browse/UIFC-360))
 * Filters SASQ does not remember the search query ([UIFC-373](https://folio-org.atlassian.net/browse/UIFC-373))
 * Fix handleNeedMoreData in FiltersRoute ([UIFC-377](https://folio-org.atlassian.net/browse/UIFC-377))
 * Fix tests and add mocks local ([UIFC-379](https://folio-org.atlassian.net/browse/UIFC-379))
 * Fix GitHub Actions workflow not running for tags ([FOLIO-4086](https://folio-org.atlassian.net/browse/FOLIO-4086))
 * Remove dependencies of k-int and stripes-erm-components ([UIFC-366](https://folio-org.atlassian.net/browse/UIFC-366))
+* Update required interfaces ([UIFC-387](https://folio-org.atlassian.net/browse/UIFC-387))
 
 ## [6.1.0](https://github.com/folio-org/ui-finc-select/tree/v6.1.0) (2024-03-20)
 * Remove deprecated pane properties ([UIFC-339](https://issues.folio.org/browse/UIFC-339))